requirements.txt
Django==5.0.6
django-ninja==1.1.0
celery==5.4.0
redis==5.0.4
psycopg2-binary==2.9.9
pgvector==0.2.5
langchain==0.2.5
langchain-community==0.2.5
python-multipart==0.0.9
twilio==9.1.0
httpx==0.27.0
python-dotenv==1.0.1
pypdf==4.2.0
sentence-transformers==2.7.0
huggingface-hub==0.23.0
groq==0.8.0
resend==2.0.0
python-docx==1.1.0

# Testing
pytest==8.0.0
pytest-django==4.8.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code quality
black==24.2.0
isort==5.13.0
flake8==7.0.0